policy: Do not specify the gettext-domain
Same fix with https://github.com/elementary/switchboard-plug-locale/commit/a407f4d83448041e05e800fd735d26664f66d136
but we embeded the same message in the plug domain thus the message in the
polkit dialog is shown as translated as expected.

However, it is better to keep the message only in the policy file for
maintainability.